Moved shared fixture to conftest.py
